create a reusable `use-draft-entity-subgraph` hook
